App Root update - update ReactDOM render invocation - add StrictMode to App - pass <Component /> to element prop for AppRoutes
