+++ src/AppRoutes.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
-import { Routes, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { Error404, Home } from '@/pages';
 
 export function AppRoutes() {
   return (
-    <Routes>
-      <Route path="/" element={Home} />
-      <Route path="*" element={Error404} />
-    </Routes>
+    <Router>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="*" element={<Error404 />} />
+      </Routes>
+    </Router>
   );
 }
